src/main/java/br/com/contato/dto/EnderecoDTO.java
package br.com.contato.dto;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.annotation.JsonProperty;

import java.util.Objects;

@JsonIgnoreProperties(ignoreUnknown = true)
public class EnderecoDTO {
    @JsonProperty("code")
    private String cep;

    @JsonProperty("address")
    private String logradouro;

    @JsonProperty("city")
    private String cidade;

    @JsonProperty("state")
    private String uf;

    @JsonProperty("district")
    private String bairro;

    @JsonProperty("status")
    private int status;

    @JsonProperty("ok")
    private boolean ok;

    @JsonProperty("statusText")
    private String statusText;

    public String getCep() {
        return cep;
    }

    public void setCep(String cep) {
        this.cep = cep;
    }

    public String getLogradouro() {
        return logradouro;
    }

    public void setLogradouro(String logradouro) {
        this.logradouro = logradouro;
    }

    public String getCidade() {
        return cidade;
    }

    public void setCidade(String cidade) {
        this.cidade = cidade;
    }

    public String getUf() {
        return uf;
    }

    public void setUf(String uf) {
        this.uf = uf;
    }

    public String getBairro() {
        return bairro;
    }

    public void setBairro(String bairro) {
        this.bairro = bairro;
    }

    public int getStatus() {
        return status;
    }

    public void setStatus(int status) {
        this.status = status;
    }

    public boolean isOk() {
        return ok;
    }

    public void setOk(boolean ok) {
        this.ok = ok;
    }

    public String getStatusText() {
        return statusText;
    }

    public void setStatusText(String statusText) {
        this.statusText = statusText;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        EnderecoDTO that = (EnderecoDTO) o;
        return Objects.equals(cep, that.cep) &&
                Objects.equals(logradouro, that.logradouro) &&
                Objects.equals(cidade, that.cidade) &&
                Objects.equals(uf, that.uf) &&
                Objects.equals(bairro, that.bairro) &&
                status == that.status &&
                ok == that.ok &&
                Objects.equals(statusText, that.statusText);
    }

    @Override
    public int hashCode() {
        return Objects.hash(cep, logradouro, cidade, uf, bairro, status, ok, statusText);
    }

}